+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `tlua::Push` trait implementations for `OsString`, `OsStr`, `Path`, `PathBuf`
 - `tlua::LuaRead` trait implementations for `OsString`, `PathBuf`
 - tlua::LuaTable::metatable which is a better alternative to the existing `tlua::LuaTable::get_or_create_metatable`
+- `ffi::tarantool::box_schema_version` and `ffi::tarantool::box_session_id` functions
 
 ### Changed